+++ src/main/java/com/raoulvdberge/refinedstorage/tile/craftingmonitor/ICraftingMonitor.java
@@ -26,5 +26,9 @@ public interface ICraftingMonitor {
 
     ItemHandlerBasic getFilter();
 
+    boolean canViewAutomated();
+
+    void onViewAutomatedChanged(boolean viewAutomated);
+
     boolean isActive();
 }